refactor(pane): split pane_tools.py into domain subpackage
pane_tools.py had grown to 1315 lines — a kitchen-sink module
covering I/O, capture, wait/poll, search, copy-mode, paste, resize,
kill, select, swap, pipe, display, and snapshot all in one file.
Large modules in this project are especially costly because each tool
is AI-surface: reviewers need to audit behavior and docstrings
together, and a flat 1.3k-line file makes the diff graph noisy when
any one tool changes.

Convert the module into a ``pane_tools/`` package with one file per
operation domain:

  io.py         send_keys, capture_pane, clear_pane, paste_text
  wait.py       wait_for_text, wait_for_content_change
  search.py     search_panes
  copy_mode.py  enter_copy_mode, exit_copy_mode
  layout.py     resize_pane, select_pane, swap_pane
  lifecycle.py  kill_pane, set_pane_title, get_pane_info
  pipe.py       pipe_pane
  meta.py       display_message, snapshot_pane

``__init__.py`` re-exports every tool by name so
``from libtmux_mcp.tools.pane_tools import send_keys`` still works,
preserving existing test imports and any downstream code that pinned
to the flat namespace. The ``register()`` function is unchanged in
behavior — all tool names, titles, annotations, and safety tags are
identical so MCP clients see no surface change.

No tool bodies were modified; this is a pure move. All 188 existing
tests pass unchanged.
